feat: refactor OTP verification page and extract Verification component
